requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.5
pypdf==4.2.0
werkzeug==3.0.1
requests==2.31.0
numpy==1.26.4
